simulator/requirements.txt
google-cloud-pubsub==2.23.0
numpy==2.1.2
pyyaml==6.0.2
reportlab==4.2.5
pytest==8.3.3
# Phase 2: PySensorMQTT — MQTT-based real hardware sensor integration
# pysensormqtt
